+++ issuer-front-end/src/Main.tsx
@@ -1,8 +1,7 @@
 /* eslint-disable no-alert */
 /* eslint-disable no-console */
 /* eslint-disable no-param-reassign */
-import React, { useEffect, useState, ChangeEvent, PropsWithChildren, MouseEvent, useCallback } from 'react';
-import { saveAs } from 'file-saver';
+import React, { useEffect, useState } from 'react';
 import {
     WalletConnectionProps,
     useConnection,
@@ -11,152 +10,14 @@ import {
     TESTNET,
     MAINNET,
 } from '@concordium/react-components';
-import { Button, Col, Row, Form, InputGroup } from 'react-bootstrap';
-import { AccountAddress, TransactionKindString, TransactionSummaryType } from '@concordium/web-sdk';
-import { TailSpin } from 'react-loader-spinner';
+
+import { AccountAddress } from '@concordium/web-sdk';
+
 import { WalletConnectionTypeButton } from './WalletConnectorTypeButton';
 
-import { createNewIssuer } from './writing_to_blockchain';
-
-import {
-    BROWSER_WALLET,
-    REFRESH_INTERVAL,
-    EXAMPLE_CREDENTIAL_SCHEMA_OBJECT,
-    EXAMPLE_ISSUER_METADATA_OBJECT,
-    EXAMPLE_CREDENTIAL_METADATA_OBJECT,
-} from './constants';
-
-type CredentialSchema = {
-    name: string;
-    description: string;
-    type: string;
-    properties: {
-        credentialSubject: {
-            type: string;
-            properties: {
-                id: {
-                    title: string;
-                    type: string;
-                    description: string;
-                };
-                attributes: {
-                    title: string;
-                    description: string;
-                    type: string;
-                    properties: object;
-                    required: string[];
-                };
-            };
-            required: string[];
-        };
-    };
-    required: string[];
-};
-
-type SchemaRef = {
-    schema_ref: {
-        hash: {
-            None: [];
-        };
-        url: string;
-    };
-};
-
-function TestBox({ children }: PropsWithChildren) {
-    return (
-        <fieldset className="testBox">
-            <div className="testBoxFields">{children}</div>
-            <br />
-        </fieldset>
-    );
-}
-
-async function addRevokationKey(
-    revocationKeys: string[],
-    setRevocationKeys: (value: string[]) => void,
-    setRevoationKeyInput: (value: string) => void,
-    newRevocationKey: string | undefined
-) {
-    if (newRevocationKey === undefined) {
-        throw new Error(`Set revocation key`);
-    }
-    if (revocationKeys.includes(newRevocationKey)) {
-        throw new Error(`Duplicate revocation key: ${newRevocationKey}`);
-    }
-    if (newRevocationKey.length !== 64) {
-        throw new Error(`Revocation key should have a length of 64`);
-    }
-    if (newRevocationKey) {
-        setRevocationKeys([...revocationKeys, newRevocationKey]);
-        setRevoationKeyInput('');
-    }
-}
-
-async function addAttribute(
-    attributes: object[],
-    setAttributes: (value: object[]) => void,
-    attributeTitle: string | undefined,
-    attributeDescription: string | undefined,
-    isRequired: boolean,
-    type: string | undefined,
-    credentialSchema: CredentialSchema
-) {
-    if (attributeTitle === undefined) {
-        throw new Error(`AttributeTitle needs to be set`);
-    }
-
-    if (attributeDescription === undefined) {
-        throw new Error(`AttributeDescription needs to be set`);
-    }
-
-    if (type === undefined) {
-        throw new Error(`Type needs to be set`);
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    attributes.forEach((value: any) => {
-        if (value[attributeTitle.replaceAll(' ', '')] !== undefined) {
-            throw new Error(`Duplicate attribute key: "${attributeTitle}"`);
-        }
-    });
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const newAttribute: any = {};
-    if (type === 'date-time') {
-        newAttribute[attributeTitle.replaceAll(' ', '')] = {
-            title: attributeTitle,
-            type: 'object',
-            description: attributeDescription,
-            properties: {
-                type: {
-                    type: 'string',
-                    const: 'date-time',
-                },
-                timestamp: {
-                    type: 'string',
-                    format: 'date-time',
-                },
-            },
-            required: ['type', 'timestamp'],
-        };
-    } else {
-        newAttribute[attributeTitle.replaceAll(' ', '')] = {
-            title: attributeTitle,
-            type,
-            description: attributeDescription,
-        };
-    }
-
-    credentialSchema.properties.credentialSubject.properties.attributes.properties = [...attributes, newAttribute];
-
-    if (isRequired) {
-        credentialSchema.properties.credentialSubject.properties.attributes.required.push(
-            attributeTitle.replaceAll(' ', '')
-        );
-    }
-
-    setAttributes([...attributes, newAttribute]);
-}
+import { BROWSER_WALLET, REFRESH_INTERVAL } from './constants';
+import CreateSchemaAndMetadataFiles from './CreateSchemaAndMetadataFiles';
+import DeployCredentialContract from './DeployCredentialContract';
 
 interface ConnectionProps {
     walletConnectionProps: WalletConnectionProps;
@@ -172,236 +33,15 @@ export default function Main(props: ConnectionProps) {
     const { connection, setConnection, account } = useConnection(connectedAccounts, genesisHashes);
     const { connect, isConnecting, connectError } = useConnect(activeConnector, setConnection);
 
-    const [credentialSchema, setCredentialSchema] = useState(EXAMPLE_CREDENTIAL_SCHEMA_OBJECT);
-    const [credentialMetadata, setCredentialMetadata] = useState(EXAMPLE_CREDENTIAL_METADATA_OBJECT);
-    const [issuerMetadata, setIssuerMetadata] = useState(EXAMPLE_ISSUER_METADATA_OBJECT);
-    const [attributes, setAttributes] = useState<object[]>([]);
-
-    const [credentialName, setCredentialName] = useState('Education certificate');
-    const [credentialDescription, setCredentialDescription] = useState(
-        'Simple representation of an education certificate.'
-    );
-
-    const [backgroundColor, setBackgroundColor] = useState('#92a8d1');
-    const [logo, setLogo] = useState('https://avatars.githubusercontent.com/u/39614219?s=200&v=4');
-    const [title, setTitle] = useState('Example Title');
-
-    const [iconURL, setIconURL] = useState('https://concordium.com/wp-content/uploads/2022/07/Concordium-1.png');
-    const [URL, setURL] = useState('https://concordium.com');
-    const [issuerDescription, setIssuerDescription] = useState('A public-layer 1, science-backed blockchain');
-    const [issuerName, setIssuerName] = useState('Concordium');
-
-    const [attributeTitle, setAttributeTitle] = useState<string | undefined>(undefined);
-    const [attributeDescription, setAttributeDescription] = useState<string | undefined>(undefined);
-
-    const [attributeType, setAttributeType] = useState<string>();
-    const [required, setRequired] = useState(false);
-
-    const [showCredentialSchema, setShowCredentialSchema] = useState(false);
-    const [showCredentialMetadata, setShowCredentialMetadata] = useState(false);
-    const [showIssuerMetadata, setShowIssuerMetadata] = useState(false);
-
-    const [smartContractIndexError, setSmartContractIndexError] = useState('');
-    const [viewErrorModuleReference, setViewErrorModuleReference] = useState('');
-    const [waitingForTransactionToFinialize, setWaitingForTransactionToFinialize] = useState(false);
-
-    const [smartContractIndex, setSmartContractIndex] = useState('');
-
     const [viewErrorAccountBalance, setViewErrorAccountBalance] = useState('');
-    const [transactionError, setTransactionError] = useState('');
-    const [userInputErrorAttributes, setUserInputErrorAttributes] = useState('');
-    const [userInputError, setUserInputError] = useState('');
 
     const [accountExistsOnNetwork, setAccountExistsOnNetwork] = useState(true);
 
     const [isWaitingForTransaction, setWaitingForUser] = useState(false);
 
-    const [issuerKey, setIssuerKey] = useState<string | undefined>(undefined);
-
     const [accountBalance, setAccountBalance] = useState('');
-    const [txHash, setTxHash] = useState('');
-
-    const [issuerMetaData, setIssuerMetaData] = useState<string | undefined>(undefined);
-
-    const [credentialType, setCredentialType] = useState<string | undefined>(undefined);
-    const [schemaCredential, setSchemaCredential] = useState<SchemaRef | undefined>(undefined);
-
-    const [revocationKeys, setRevocationKeys] = useState<string[]>([]);
-    const [revocationKeyInput, setRevocationKeyInput] = useState<string | undefined>(undefined);
-
-    const display = useCallback((event: MouseEvent<HTMLElement>) => {
-        const element = event.target as HTMLTextAreaElement;
-        alert(element.parentElement?.title || element.parentElement?.parentElement?.title || element.title);
-    }, []);
-
-    const changeDropDownHandler = () => {
-        const e = document.getElementById('write') as HTMLSelectElement;
-        const sel = e.selectedIndex;
-        const { value } = e.options[sel];
-        setAttributeType(value);
-    };
-
-    const changeAttributeDescription = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setAttributeDescription(target.value);
-    }, []);
-
-    const changeAttributeTitle = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setAttributeTitle(target.value);
-    }, []);
-
-    const changeIssuerMetaDataURLHandler = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setIssuerMetaData(target.value);
-    }, []);
-
-    const changeCheckBox = useCallback((requiredValue: boolean, event: ChangeEvent) => {
-        const target = event.target as HTMLInputElement;
-        target.checked = !requiredValue;
-
-        setRequired(!requiredValue);
-    }, []);
-
-    const changeCredentialDescription = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setCredentialDescription(target.value);
-
-        const newCredentialSchema = credentialSchema;
-        newCredentialSchema.description = target.value;
-        setCredentialSchema(newCredentialSchema);
-    }, []);
-
-    const changeCredentialName = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setCredentialName(target.value);
-
-        const newCredentialSchema = credentialSchema;
-        newCredentialSchema.name = target.value;
-        setCredentialSchema(newCredentialSchema);
-    }, []);
-
-    const changeBackgroundColor = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setBackgroundColor(target.value);
-
-        const newCredentialMetadata = credentialMetadata;
-        newCredentialMetadata.backgroundColor = target.value;
-        setCredentialMetadata(newCredentialMetadata);
-    }, []);
-
-    const changeTitle = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setTitle(target.value);
-
-        const newCredentialMetadata = credentialMetadata;
-        newCredentialMetadata.title = target.value;
-        setCredentialMetadata(newCredentialMetadata);
-    }, []);
-
-    const changeLogoURL = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setLogo(target.value);
-
-        const newCredentialMetadata = credentialMetadata;
-        newCredentialMetadata.logo.url = target.value;
-        setCredentialMetadata(newCredentialMetadata);
-    }, []);
-
-    const changeIconURL = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setIconURL(target.value);
-
-        const newIssuerMetadata = issuerMetadata;
-        newIssuerMetadata.icon.url = target.value;
-        setIssuerMetadata(newIssuerMetadata);
-    }, []);
-
-    const changeURL = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setURL(target.value);
-
-        const newIssuerMetadata = issuerMetadata;
-        newIssuerMetadata.url = target.value;
-        setIssuerMetadata(newIssuerMetadata);
-    }, []);
-
-    const changeIssuerDescription = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setIssuerDescription(target.value);
-
-        const newIssuerMetadata = issuerMetadata;
-        newIssuerMetadata.description = target.value;
-        setIssuerMetadata(newIssuerMetadata);
-    }, []);
-
-    const changeIssuerName = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setIssuerName(target.value);
-
-        const newIssuerMetadata = issuerMetadata;
-        newIssuerMetadata.name = target.value;
-        setIssuerMetadata(newIssuerMetadata);
-    }, []);
-
-    const changeIssuerKeyHandler = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setIssuerKey(target.value);
-    }, []);
 
     const client = useGrpcClient(isTestnet ? TESTNET : MAINNET);
-
-    const changeCredentialSchemaURLHandler = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setSchemaCredential({
-            schema_ref: {
-                hash: {
-                    None: [],
-                },
-                url: target.value,
-            },
-        });
-    }, []);
-
-    const changeCredentialTypeHandler = useCallback((event: ChangeEvent) => {
-        const target = event.target as HTMLTextAreaElement;
-        setCredentialType(target.value);
-    }, []);
-
-    // Refresh smartContractIndex periodically.
-    // eslint-disable-next-line consistent-return
-    useEffect(() => {
-        if (connection && client && account && txHash !== '') {
-            const interval = setInterval(() => {
-                console.log('refreshing_smartContractIndex');
-                client
-                    .getBlockItemStatus(txHash)
-                    .then((report) => {
-                        if (report !== undefined) {
-                            setViewErrorModuleReference('');
-                            if (report.status === 'finalized') {
-                                setWaitingForTransactionToFinialize(false);
-                                if (
-                                    report.outcome.summary.type === TransactionSummaryType.AccountTransaction &&
-                                    report.outcome.summary.transactionType === TransactionKindString.InitContract
-                                ) {
-                                    setSmartContractIndexError('');
-                                    setSmartContractIndex(
-                                        report.outcome.summary.contractInitialized.address.index.toString()
-                                    );
-                                } else {
-                                    setSmartContractIndexError('Contract initialization failed');
-                                }
-                            }
-                        }
-                    })
-                    .catch((e) => {
-                        setViewErrorModuleReference((e as Error).message);
-                    });
-            }, REFRESH_INTERVAL.asMilliseconds());
-            return () => clearInterval(interval);
-        }
-    }, [connection, account, client, txHash]);
 
     // Refresh accountInfo periodically.
     // eslint-disable-next-line consistent-return
@@ -522,598 +162,15 @@ export default function Main(props: ConnectionProps) {
                                     Error: {viewErrorAccountBalance}.
                                 </div>
                             )}
-                            {active === 3 && (
-                                <>
-                                    <TestBox>
-                                        <div
-                                            className="containerToolTip"
-                                            role="presentation"
-                                            onClick={display}
-                                            data-toggle="tooltip"
-                                            title="The credentialSchema is a JSON object that you will create in this step. You should host this JSON object somewhere publicly on the web so it is available via a URL. You will use the URL in step 4. You can for example host it on your gist account (`https://gist.github.com/`) and click the `raw` button to optain the URL (e.g. `https://gist.githubusercontent.com/DOBEN/bfe30ecea16f7a3ea1b87aa40902b9ac/raw/a8ab51fca489d04710fb19fb7122bb283dba719a/gistfile1.txt`)."
-                                        >
-                                            <div>
-                                                <h3>CredentialSchema</h3>
-                                            </div>
-                                            <div className="infolink" />
-                                        </div>
-                                        <br />
-                                        <br />
-                                        Add <strong>CredentialName</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="issuerKey"
-                                            type="text"
-                                            value={credentialName}
-                                            onChange={changeCredentialName}
-                                        />
-                                        <br />
-                                        <br />
-                                        Add <strong>CredentialDescription</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="credentialDescription"
-                                            type="text"
-                                            value={credentialDescription}
-                                            onChange={changeCredentialDescription}
-                                        />
-                                        <br />
-                                        <br />
-                                        <TestBox>
-                                            Add <strong>AttributeTitle</strong>:
-                                            <br />
-                                            <input
-                                                className="inputFieldStyle"
-                                                id="attributeTitle"
-                                                type="text"
-                                                value={attributeTitle}
-                                                onChange={changeAttributeTitle}
-                                            />
-                                            <br />
-                                            <br />
-                                            Add <strong>AttributeDescription</strong>:
-                                            <br />
-                                            <input
-                                                className="inputFieldStyle"
-                                                id="attributeDescription"
-                                                type="text"
-                                                value={attributeDescription}
-                                                onChange={changeAttributeDescription}
-                                            />
-                                            <label className="field">
-                                                Select Type:
-                                                <br />
-                                                <br />
-                                                <select name="write" id="write" onChange={changeDropDownHandler}>
-                                                    <option value="choose" disabled selected>
-                                                        Choose
-                                                    </option>
-                                                    <option value="integer">Integer</option>
-                                                    <option value="string">String</option>
-                                                    <option value="date-time">DateTime</option>
-                                                </select>
-                                            </label>
-                                            <br />
-                                            <br />
-                                            <div>
-                                                <input
-                                                    type="checkbox"
-                                                    id="checkBox"
-                                                    name="checkBox"
-                                                    onChange={(event) => changeCheckBox(required, event)}
-                                                />
-                                                <label htmlFor="checkBox">&nbsp;Is Type Required</label>
-                                            </div>
-                                            <br />
-                                            <br />
-                                            <button
-                                                className="btn btn-primary"
-                                                type="button"
-                                                onClick={() => {
-                                                    setUserInputErrorAttributes('');
-                                                    addAttribute(
-                                                        attributes,
-                                                        setAttributes,
-                                                        attributeTitle,
-                                                        attributeDescription,
-                                                        required,
-                                                        attributeType,
-                                                        credentialSchema
-                                                    ).catch((err: Error) =>
-                                                        setUserInputErrorAttributes((err as Error).message)
-                                                    );
-                                                }}
-                                            >
-                                                Add Attribute
-                                            </button>
-                                            <button
-                                                className="btn btn-primary"
-                                                type="button"
-                                                onClick={() => {
-                                                    setAttributes([]);
-                                                    setAttributeTitle('');
-                                                    setAttributeDescription('');
-                                                    setAttributeType(undefined);
-                                                    setUserInputErrorAttributes('');
-                                                }}
-                                            >
-                                                Clear All Attributes
-                                            </button>
-                                            <br />
-                                            {attributes.length !== 0 && (
-                                                <>
-                                                    <div className="actionResultBox">
-                                                        <div>
-                                                            You have added the following <strong>attributes</strong>:
-                                                        </div>
-                                                        <div>
-                                                            <pre className="largeText">
-                                                                {JSON.stringify(attributes, null, '\t')}
-                                                            </pre>
-                                                        </div>
-                                                    </div>
-                                                    <br />
-                                                    <br />
-                                                </>
-                                            )}
-                                            {userInputErrorAttributes !== '' && (
-                                                <div className="alert alert-danger" role="alert">
-                                                    Error: {userInputErrorAttributes}.
-                                                </div>
-                                            )}
-                                        </TestBox>
-                                        <br />
-                                        <br />
-                                        <button
-                                            className="btn btn-primary"
-                                            type="button"
-                                            onClick={() => {
-                                                setShowCredentialSchema(true);
-                                            }}
-                                        >
-                                            Create CredentialSchema
-                                        </button>
-                                        <button
-                                            className="btn btn-primary"
-                                            type="button"
-                                            onClick={() => {
-                                                setShowCredentialSchema(true);
-
-                                                const fileName = 'credentialSchema.json';
-
-                                                const fileToSave = new Blob([JSON.stringify(credentialSchema)], {
-                                                    type: 'application/json',
-                                                });
-
-                                                saveAs(fileToSave, fileName);
-                                            }}
-                                        >
-                                            Download CredentialSchema
-                                        </button>
-                                        {showCredentialSchema && (
-                                            <pre className="largeText">
-                                                {JSON.stringify(credentialSchema, null, '\t')}
-                                            </pre>
-                                        )}
-                                    </TestBox>
-                                    <TestBox>
-                                        <div
-                                            className="containerToolTip"
-                                            role="presentation"
-                                            onClick={display}
-                                            data-toggle="tooltip"
-                                            title="The credentialMetadata is a JSON object that you will create in this step. You should host this JSON object somewhere publicly on the web so it is available via a URL. You will use the URL when issuing credentials. You can for example host it on your gist account (`https://gist.github.com/`) and click the `raw` button to optain the URL (e.g. `https://gist.githubusercontent.com/abizjak/ff1e90d82c5446c0e001ee6d4e33ea6b/raw/4528363aff42e3ff36b50a1d873287f2f520d610/metadata.json`)."
-                                        >
-                                            <div>
-                                                <h3>CredentialMetadata</h3>
-                                            </div>
-                                            <div className="infolink" />
-                                        </div>
-                                        <br />
-                                        <br />
-                                        Add <strong>Title</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="title"
-                                            type="text"
-                                            value={title}
-                                            onChange={changeTitle}
-                                        />
-                                        <br />
-                                        <br />
-                                        Add <strong>LogoURL</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="logoURL"
-                                            type="text"
-                                            value={logo}
-                                            onChange={changeLogoURL}
-                                        />{' '}
-                                        <br />
-                                        <br />
-                                        Add <strong>BackGroundColor</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="backgroundColor"
-                                            type="text"
-                                            value={backgroundColor}
-                                            onChange={changeBackgroundColor}
-                                        />
-                                        <button
-                                            className="btn btn-primary"
-                                            type="button"
-                                            onClick={() => {
-                                                setShowCredentialMetadata(true);
-                                            }}
-                                        >
-                                            Create CredentialMetadata
-                                        </button>
-                                        <button
-                                            className="btn btn-primary"
-                                            type="button"
-                                            onClick={() => {
-                                                setShowCredentialMetadata(true);
-
-                                                const fileName = 'credentialMetadata.json';
-
-                                                const fileToSave = new Blob([JSON.stringify(credentialMetadata)], {
-                                                    type: 'application/json',
-                                                });
-
-                                                saveAs(fileToSave, fileName);
-                                            }}
-                                        >
-                                            Download CredentialMetadata
-                                        </button>
-                                        {showCredentialMetadata && (
-                                            <pre className="largeText">
-                                                {JSON.stringify(credentialMetadata, null, '\t')}
-                                            </pre>
-                                        )}
-                                    </TestBox>
-                                    <TestBox>
-                                        <div
-                                            className="containerToolTip"
-                                            role="presentation"
-                                            onClick={display}
-                                            data-toggle="tooltip"
-                                            title="The issuerMetadata is a JSON object that you will create in this step. You should host this JSON object somewhere publicly on the web so it is available via a URL. You will use the URL in step 4. You can for example host it on your gist account (`https://gist.github.com/`) and click the `raw` button to optain the URL (e.g. `https://gist.githubusercontent.com/DOBEN/d12deee42e06601efb72859da9be5759/raw/137a9a4b9623dfe16fa8e9bb7ab07f5858d92c53/gistfile1.txt`)."
-                                        >
-                                            <div>
-                                                <h3>IssuerMetadata</h3>
-                                            </div>
-                                            <div className="infolink" />
-                                        </div>
-                                        <br />
-                                        <br />
-                                        Add <strong>IssuerName</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="issuerName"
-                                            type="text"
-                                            value={issuerName}
-                                            onChange={changeIssuerName}
-                                        />
-                                        <br />
-                                        <br />
-                                        Add <strong>IssuerDescription</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="issuerDescription"
-                                            type="text"
-                                            value={issuerDescription}
-                                            onChange={changeIssuerDescription}
-                                        />
-                                        Add <strong>URL</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="URL"
-                                            type="text"
-                                            value={URL}
-                                            onChange={changeURL}
-                                        />
-                                        <br />
-                                        <br />
-                                        Add <strong>IconURL</strong>:
-                                        <br />
-                                        <input
-                                            className="inputFieldStyle"
-                                            id="iconURL"
-                                            type="text"
-                                            value={iconURL}
-                                            onChange={changeIconURL}
-                                        />
-                                        <button
-                                            className="btn btn-primary"
-                                            type="button"
-                                            onClick={() => {
-                                                setShowIssuerMetadata(true);
-                                            }}
-                                        >
-                                            Create IssuerMetadata
-                                        </button>
-                                        <button
-                                            className="btn btn-primary"
-                                            type="button"
-                                            onClick={() => {
-                                                setShowIssuerMetadata(true);
-
-                                                const fileName = 'issuerMetadata.json';
-
-                                                const fileToSave = new Blob([JSON.stringify(issuerMetadata)], {
-                                                    type: 'application/json',
-                                                });
-
-                                                saveAs(fileToSave, fileName);
-                                            }}
-                                        >
-                                            Download IssuerMetadata
-                                        </button>
-                                        {showIssuerMetadata && (
-                                            <pre className="largeText">
-                                                {JSON.stringify(issuerMetadata, null, '\t')}
-                                            </pre>
-                                        )}
-                                    </TestBox>
-                                </>
-                            )}
+                            {/* Step 3: Create schema and metadata files */}
+                            {active === 3 && <CreateSchemaAndMetadataFiles />}
+                            {/* Step 4: Deploy issuer smart contract */}
                             {active === 4 && (
-                                <TestBox>
-                                    <div
-                                        className="containerToolTip"
-                                        role="presentation"
-                                        onClick={display}
-                                        data-toggle="tooltip"
-                                        title="If you become an issuer, you will need to sign the credentials with your issuer private key at the backend before a credential that you issue can be added to a holder's wallet. For testing purposes on testnet, you can create a public-private key pair with an online tool (e.g. https://cyphr.me/ed25519_tool/ed.html) and use the public key as the isserKey here. The issuerKey should have a length of 64 characters e.g. `8fe0dc02ffbab8d30410233ed58b44a53c418b368ae91cdcdbcdb9e79358be82`."
-                                    >
-                                        <div>
-                                            Add <strong>IssuerKey</strong>
-                                        </div>
-                                        <div className="infolink" />
-                                        &nbsp;:
-                                    </div>
-                                    <br />
-                                    <input
-                                        className="inputFieldStyle"
-                                        id="issuerKey"
-                                        type="text"
-                                        value={issuerKey}
-                                        onChange={changeIssuerKeyHandler}
-                                    />
-                                    <br />
-                                    <br />
-                                    <div
-                                        className="containerToolTip"
-                                        role="presentation"
-                                        onClick={display}
-                                        data-toggle="tooltip"
-                                        title="The issuerMetadata file that you created in the previous step should be hosted on the web. You can for example host it on your gist account (`https://gist.github.com/`) and click the `raw` button to optain the URL. You should then input the URL where you host your issuerMetadata here (e.g. `https://gist.githubusercontent.com/DOBEN/d12deee42e06601efb72859da9be5759/raw/137a9a4b9623dfe16fa8e9bb7ab07f5858d92c53/gistfile1.txt`)."
-                                    >
-                                        <div>
-                                            Add <strong>IssuerMetadata</strong>
-                                        </div>
-                                        <div className="infolink" />
-                                        &nbsp;:
-                                    </div>
-                                    <br />
-                                    <input
-                                        className="inputFieldStyle"
-                                        id="issuerMetaDataURL"
-                                        type="text"
-                                        value={issuerMetaData}
-                                        onChange={changeIssuerMetaDataURLHandler}
-                                    />
-                                    <br />
-                                    <br />
-                                    <div
-                                        className="containerToolTip"
-                                        role="presentation"
-                                        onClick={display}
-                                        data-toggle="tooltip"
-                                        title="You should define a type for your credential (e.g. `myCredentialType` or `EducationalCertificate`)"
-                                    >
-                                        <div>
-                                            Add <strong>CredentialType</strong>
-                                        </div>
-                                        <div className="infolink" />
-                                        &nbsp;:
-                                    </div>
-                                    <br />
-                                    <input
-                                        className="inputFieldStyle"
-                                        id="credentialType"
-                                        type="text"
-                                        value={credentialType}
-                                        onChange={changeCredentialTypeHandler}
-                                    />
-                                    <br />
-                                    <br />
-                                    <div
-                                        className="containerToolTip"
-                                        role="presentation"
-                                        onClick={display}
-                                        data-toggle="tooltip"
-                                        title="The credentialSchema file that you created in the previous step should be hosted on the web. You can for example host it on your gist account (`https://gist.github.com/`) and click the `raw` button to optain the URL. You should then input the URL where you host your credentialSchema here (e.g. `https://gist.githubusercontent.com/DOBEN/bfe30ecea16f7a3ea1b87aa40902b9ac/raw/a8ab51fca489d04710fb19fb7122bb283dba719a/gistfile1.txt`)."
-                                    >
-                                        <div>
-                                            Add <strong>CredentialSchema</strong>
-                                        </div>
-                                        <div className="infolink" />
-                                        &nbsp;:
-                                    </div>
-                                    <br />
-                                    <input
-                                        className="inputFieldStyle"
-                                        id="credentialSchemaURL"
-                                        type="text"
-                                        value={schemaCredential?.schema_ref.url}
-                                        onChange={changeCredentialSchemaURLHandler}
-                                    />
-                                    <br />
-                                    <br />
-                                    {revocationKeys.length !== 0 && (
-                                        <>
-                                            <div className="actionResultBox">
-                                                <div>
-                                                    You have added the following <strong>revocationKeys</strong>:
-                                                </div>
-                                                <div>
-                                                    {revocationKeys?.map((element) => (
-                                                        <li key={element}>{element}</li>
-                                                    ))}
-                                                </div>
-                                            </div>
-                                            <br />
-                                            <br />
-                                        </>
-                                    )}
-                                    {userInputError !== '' && (
-                                        <div className="alert alert-danger" role="alert">
-                                            Error: {userInputError}.
-                                        </div>
-                                    )}
-                                    <Form
-                                        onSubmit={(e) => {
-                                            e.preventDefault();
-                                            setUserInputError('');
-                                            addRevokationKey(
-                                                revocationKeys,
-                                                setRevocationKeys,
-                                                setRevocationKeyInput,
-                                                revocationKeyInput
-                                            ).catch((err: Error) => setUserInputError((err as Error).message));
-                                        }}
-                                    >
-                                        <div
-                                            className="containerToolTip"
-                                            role="presentation"
-                                            onClick={display}
-                                            data-toggle="tooltip"
-                                            title="The keys inserted here can revoke any credential that you issue. You can leave this an empty array if you don't want to grant such permissions to special revocation keys. For testing purposes on testnet, you can create public-private key pairs with an online tool (e.g. https://cyphr.me/ed25519_tool/ed.html) and use the public keys here. Each revocationKey should have a length of 64 characters e.g. `8fe0dc02ffbab8d30410233ed58b44a53c418b368ae91cdcdbcdb9e79358be82`."
-                                        >
-                                            <div>
-                                                Add <strong>RevocationKeys</strong>
-                                            </div>
-                                            <div className="infolink" />
-                                            &nbsp;:
-                                        </div>
-                                        <br />
-                                        <Row>
-                                            <Col sm={10}>
-                                                <InputGroup className="mb-3">
-                                                    <Form.Control
-                                                        value={revocationKeyInput}
-                                                        onChange={(e) => setRevocationKeyInput(e.target.value)}
-                                                    />
-                                                    <Button type="submit" variant="outline-secondary">
-                                                        Add
-                                                    </Button>
-                                                </InputGroup>
-                                            </Col>
-                                            <Col sm={1}>
-                                                <Button
-                                                    variant="outline-secondary"
-                                                    onClick={() => {
-                                                        setRevocationKeys([]);
-                                                        setRevocationKeyInput('');
-                                                        setUserInputError('');
-                                                    }}
-                                                >
-                                                    Clear
-                                                </Button>
-                                            </Col>
-                                        </Row>
-                                    </Form>
-                                    <button
-                                        className="btn btn-primary"
-                                        type="button"
-                                        onClick={() => {
-                                            setTxHash('');
-                                            setTransactionError('');
-                                            setSmartContractIndex('');
-                                            setWaitingForTransactionToFinialize(true);
-
-                                            const tx = createNewIssuer(
-                                                connection,
-                                                account,
-                                                issuerMetaData,
-                                                issuerKey,
-                                                schemaCredential,
-                                                JSON.stringify(revocationKeys),
-                                                credentialType
-                                            );
-                                            tx.then(setTxHash).catch((err: Error) => {
-                                                setTransactionError((err as Error).message);
-                                                setWaitingForTransactionToFinialize(false);
-                                            });
-                                        }}
-                                    >
-                                        Create New Issuer
-                                    </button>
-                                    <br />
-                                    <br />
-                                    {!txHash && transactionError && (
-                                        <div className="alert alert-danger" role="alert">
-                                            Error: {transactionError}.
-                                        </div>
-                                    )}
-                                    {smartContractIndexError !== '' && (
-                                        <div className="alert alert-danger" role="alert">
-                                            Error: {smartContractIndexError}.
-                                        </div>
-                                    )}
-                                    {viewErrorModuleReference && (
-                                        <div className="alert alert-danger" role="alert">
-                                            Error: {viewErrorModuleReference}.
-                                        </div>
-                                    )}
-                                    {txHash && (
-                                        <div>
-                                            <div>Transaction hash:</div>
-                                            <a
-                                                className="link"
-                                                target="_blank"
-                                                rel="noreferrer"
-                                                href={`https://${
-                                                    isTestnet ? `testnet.` : ``
-                                                }ccdscan.io/?dcount=1&dentity=transaction&dhash=${txHash}`}
-                                            >
-                                                {txHash}
-                                            </a>
-                                        </div>
-                                    )}
-                                    <br />
-                                    <br />
-                                    {waitingForTransactionToFinialize === true && (
-                                        <div className="containerTwoItems">
-                                            <TailSpin
-                                                height="30"
-                                                width="30"
-                                                color="#308274"
-                                                ariaLabel="tail-spin-loading"
-                                                radius="1"
-                                                wrapperStyle={{}}
-                                                wrapperClass=""
-                                                visible
-                                            />
-                                            <div>Waiting for transaction to finalize</div>
-                                        </div>
-                                    )}
-                                    {smartContractIndex !== '' && (
-                                        <div className="actionResultBox">
-                                            Smart Contract Index:
-                                            <div>{smartContractIndex}</div>
-                                        </div>
-                                    )}
-                                </TestBox>
+                                <DeployCredentialContract
+                                    account={account}
+                                    isTestnet={isTestnet}
+                                    connection={connection}
+                                />
                             )}
                         </div>
                     )}
